% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_cmdstan.R
\name{extract_params_g0}
\alias{extract_params_g0}
\title{Extract results from cmdstanr (LNM-1.0 with gamma=0) and return data.frame with the parameters which are relevant for downstream analysis
Adapted routine for a statistical model where gamma is 0, the corresponding statistical model is not part of this repository.}
\usage{
extract_params_g0(
  param_df,
  sig = 0.05,
  nThread = 1,
  delta_thr = 0.02,
  pi_thr = 1e-04,
  hl_thr = 0.05,
  hl_thr2 = 0.1
)
}
\description{
Extract results from cmdstanr (LNM-1.0 with gamma=0) and return data.frame with the parameters which are relevant for downstream analysis
Adapted routine for a statistical model where gamma is 0, the corresponding statistical model is not part of this repository.
}
