% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_cmdstan.R
\name{extract_params_nb}
\alias{extract_params_nb}
\title{Extract results from cmdstanr (NB-1.0) and return data.frame with the parameters which are relevant for downstream analysis
NB-1.0 is not part of the github repository}
\usage{
extract_params_nb(
  param_df,
  sig = 0.05,
  nThread = 1,
  delta_thr = 0.02,
  pi_thr = 1e-04,
  hl_thr = 0.05,
  hl_thr2 = 0.1
)
}
\description{
Extract results from cmdstanr (NB-1.0) and return data.frame with the parameters which are relevant for downstream analysis
NB-1.0 is not part of the github repository
}
