% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_fun.R
\name{get_norm_facs}
\alias{get_norm_facs}
\title{remove genes/ercc spike-ins with very low/no counts, return normTMM}
\usage{
get_norm_facs(rif_reads, ercc_reads, count_min = 10, sample_min = 3)
}
\description{
remove genes/ercc spike-ins with very low/no counts, return normTMM
}
