% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_fun.R
\name{get_norm_facs_nonzero}
\alias{get_norm_facs_nonzero}
\title{remove genes/ercc spike-ins with very low/no counts, return normTMM not using zero counts}
\usage{
get_norm_facs_nonzero(rif_reads, ercc_reads, count_min = 10, sample_min = 3)
}
\description{
remove genes/ercc spike-ins with very low/no counts, return normTMM not using zero counts
}
