% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_fun.R
\name{normEXP}
\alias{normEXP}
\title{This function is similar to normTMM, but it uses the natural logarithm instead of log2.
In practice, the results for the normalization factors hardly depend on the choice of basis.}
\usage{
normEXP(ercc_reads, N)
}
\description{
This function is similar to normTMM, but it uses the natural logarithm instead of log2.
In practice, the results for the normalization factors hardly depend on the choice of basis.
}
