% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_fun.R
\name{normTMMr}
\alias{normTMMr}
\title{This normalization method is similar to normTMM, but it uses a reference sample
instead of using the geometric mean as reference.}
\usage{
normTMMr(ercc_reads, N, nref)
}
\description{
This normalization method is similar to normTMM, but it uses a reference sample
instead of using the geometric mean as reference.
}
