% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{p_value}
\alias{p_value}
\title{Fraction of samples that fall into the interval [-int0, int0], which corresponds to the null hypothesis H0.
The p value corresponds to the probability of the null hypothesis.}
\usage{
p_value(x, int0 = 0, h0 = 0)
}
\description{
Fraction of samples that fall into the interval [-int0, int0], which corresponds to the null hypothesis H0.
The p value corresponds to the probability of the null hypothesis.
}
