% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_gene}
\alias{plot_gene}
\title{Gene-wise decay plots: fitted curve + data points}
\usage{
plot_gene(
  data,
  results,
  lab,
  res_dir,
  addNorm = F,
  addDect = F,
  rm.cond = NULL,
  show.int = T,
  col = viridis(4)
)
}
\description{
Gene-wise decay plots: fitted curve + data points
}
