% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_genes}
\alias{plot_genes}
\title{gene-wise plots of (exponential) decay}
\usage{
plot_genes(
  res_dir,
  dfit,
  param,
  id_RNA = NULL,
  model_id = "",
  addNorm = F,
  addDect = F,
  rm.cond = NULL,
  show.int = T,
  col = viridis(4)
)
}
\description{
gene-wise plots of (exponential) decay
}
