% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_fun.R
\name{plot_lib_size}
\alias{plot_lib_size}
\title{plot library size}
\usage{
plot_lib_size(
  plot_name,
  dfit,
  adj = 1,
  labels = c("WT", expression(paste(Delta, "proQ")), "proQ+")
)
}
\description{
plot library size
}
