% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rpatas-package.R
\docType{package}
\name{RIF-seq-package}
\alias{RIF-seq}
\alias{RIF-seq-package}
\title{RIF-seq: R Package for the analysis of HTS RNA-decay data with Stan}
\description{
Analyse NGS sequencing time-series data with multivariate models using the statistical modeling language Stan. Stan models are in the directory stan_models, a working example can be found in the directory examples.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/BarquistLab/RIF-seq_repo}
}

}
\author{
\strong{Maintainer}: Laura Jenniches \email{laura.jenniches@helmholtz-hiri.de}

}
\keyword{internal}
