% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rpatas-package.R
\docType{package}
\name{Rpatas-package}
\alias{Rpatas}
\alias{Rpatas-package}
\title{Rpatas: R PAckage for Time-series Analysis with Stan}
\description{
Analyse NGS sequencing time-series data with multivariate models using the statistical modeling language Stan. Stan models are in the directory stan_models, example scripts and scripts which require more resources to run can be found in the directory scripts.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/BarquistLab/Rpatas}
}

}
\author{
\strong{Maintainer}: Laura Jenniches \email{laura.jenniches@helmholtz-hiri.de}

}
\keyword{internal}
