% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_fun.R
\name{logcounts}
\alias{logcounts}
\title{logcpm values, required for normalization,
otherwise, the natural logarithm is used
throughout the analysis.}
\usage{
logcounts(Y, N)
}
\description{
logcpm values, required for normalization,
otherwise, the natural logarithm is used
throughout the analysis.
}
